Generate random coordinates to place a ship, have the amount of ships occupy a coordinate, check the syntax of a guess to be valid or not
create a battleship game with a 3 by 3 board and two ships and stores guess that are different. Create a function the returns coordinates that fall within a through c and 1 through 3, by starting the return with rows that is one of the letters within the letters array that only includes a b and c. The placeShips function continues to push random coordinates generated from the getRandomCoordinates function until the both ships have a coordinate to occupy