Set up simple function logic
First checks if guess has been guessed already in first if else statement, if the guess is already in the guesses set it will print a message confirming that they have already guessed this. If not, the guess is added to the guesses set; The next if else statement checks if the guess hits a ship or not. If the array that has the coordinates of the placed ships includes the guess that is being passed through the function it will print a statement confirming that a ship has been hit. Or if the guess is not included in the ships coordinates array then a message will print confirming that they have missed.