Adjusting the amount of remaining ships
Reduce the amount of ships after a guess has sunk a ship. Adjust printed message to display the amount of ships that are left after a ship has been sunk. Then checks if the amount of ships is 1 then use singular form of ship but if it becomes 0 / something other than 1 it will add the s after the word ship to be grammatically correct.