Starting game
Start game by using readlineSync to have any input pressed by a valid input to start the game. Then call place ships the randomly place the ships to be sunk. Use readlineSynce question to ask for a guess to be inputted. Use isValidGuess to check the inputted guess and returns correlating statements. Then create an object of the game and call the StartGame method to start interacting with the program.